Added a cleanup() method
This method deletes the user/group if they exist, before creating them.
This is useful while testing on a real okta org in which, if the entity already exists, the create method will fail.
The entities may not be cleaned up due to a previously failing test
Also, this cleanup is not needed while testing on the test SDK server (mock server)