Simply timeout logic by relying on node-fetch
My previous attempt would reject our own promise that wraps the fetch call, but it did not actually abort the request.

Node fetch has it’s own setTimout() implementation which will abort the network request:

https://github.com/bitinn/node-fetch/blob/4068bed356396518e7e9c5cc2b8ccd4c137c2749/src/index.js#L57

This commit leverages that library to manage the aborting of the request and simplify our side by not wrapping in an extra promise.